Initial conversion to typescript and electron-webpack.
