+++ electron-app/icons/icon.icns
