+++ electron-app/icons/icon.ico
