+++ electron-webpack.json
@@ -1,10 +1,10 @@
 {
-  "commonSourceDirectory": "common",
   "title": "MyEtherWallet",
   "main": {
-    "sourceDirectory": "electron/main"
+    "sourceDirectory": "electron-app/main",
+    "extraEntries": ["@/preload.ts"]
   },
   "renderer": {
-    "sourceDirectory": "electron/renderer"
+    "sourceDirectory": null
   }
 }
