+++ package.json
@@ -81,6 +81,8 @@
     "electron": "1.7.10",
     "electron-builder": "19.52.1",
     "electron-packager": "10.1.1",
+    "electron-webpack": "1.11.0",
+    "electron-webpack-ts": "1.2.0",
     "empty": "0.10.1",
     "enzyme": "3.3.0",
     "enzyme-adapter-react-16": "1.1.1",
@@ -113,6 +115,7 @@
     "resolve-url-loader": "2.2.1",
     "rimraf": "2.6.2",
     "sass-loader": "6.0.6",
+    "source-map-support": "0.5.2",
     "style-loader": "0.19.1",
     "thread-loader": "1.1.2",
     "ts-jest": "22.0.1",
@@ -140,10 +143,10 @@
     "prebuild": "check-node-version --package",
     "build:downloadable": "webpack --config webpack_config/webpack.html.js",
     "prebuild:downloadable": "check-node-version --package",
-    "build:electron": "webpack --config webpack_config/webpack.electron.js",
-    "build:electron:osx": "ELECTRON_OS=osx webpack --config webpack_config/webpack.electron.js",
-    "build:electron:windows": "ELECTRON_OS=windows webpack --config webpack_config/webpack.electron.js",
-    "build:electron:linux": "ELECTRON_OS=linux webpack --config webpack_config/webpack.electron.js",
+    "build:electron": "webpack --config webpack_config/webpack.electron.js && node webpack_config/buildElectron.js",
+    "build:electron:osx": "webpack --config webpack_config/webpack.electron.js && ELECTRON_OS=osx node webpack_config/buildElectron.js",
+    "build:electron:windows": "webpack --config webpack_config/webpack.electron.js && ELECTRON_OS=windows node webpack_config/buildElectron.js",
+    "build:electron:linux": "webpack --config webpack_config/webpack.electron.js && ELECTRON_OS=linux node webpack_config/buildElectron.js",
     "prebuild:electron": "check-node-version --package",
     "test:coverage": "jest --config=jest_config/jest.config.json --coverage",
     "test": "jest --config=jest_config/jest.config.json",
@@ -155,8 +158,8 @@
     "predev": "check-node-version --package",
     "dev:https": "HTTPS=true node webpack_config/devServer.js",
     "predev:https": "check-node-version --package",
-    "dev:electron": "concurrently --names 'webpack,electron' 'BUILD_ELECTRON=true node webpack_config/devServer.js' 'electron electron/main.js'",
-    "dev:electron:https": "concurrently --names 'webpack,electron' 'BUILD_ELECTRON=true HTTPS=true node webpack_config/devServer.js' 'electron electron/main.js'",
+    "dev:electron": "concurrently --kill-others --names 'webpack,electron' 'BUILD_ELECTRON=true node webpack_config/devServer.js' 'electron-webpack dev'",
+    "dev:electron:https": "concurrently --kill-others --names 'webpack,electron' 'BUILD_ELECTRON=true HTTPS=true node webpack_config/devServer.js' 'HTTPS=true electron-webpack dev'",
     "tslint": "tslint --project . --exclude common/vendor/**/*",
     "tscheck": "tsc --noEmit",
     "start": "npm run dev",
