+++ shared/electronEvents.ts
@@ -1,0 +1,12 @@
+export default {
+  UPDATE: {
+    CHECKING_FOR_UPDATE: 'UPDATE:checking-for-update',
+    UPDATE_NOT_AVAILABLE: 'UPDATE:update-not-available',
+    UPDATE_AVAILABLE: 'UPDATE:update-available',
+    DOWNLOAD_PROGRESS: 'UPDATE:download-progress',
+    UPDATE_DOWNLOADED: 'UPDATE:update-downloaded',
+    ERROR: 'UPDATE:error',
+    DOWNLOAD_UPDATE: 'UPDATE:download-update',
+    QUIT_AND_INSTALL: 'UPDATE:quit-and-install'
+  }
+};
